Struct autodiff::forward_autodiff::F

#[repr(C)]
pub struct F<V, D> {
    pub x: V,
    pub dx: D,
}

A generic forward differentiation Dual number.

The derivative is generic in V to support composition and alternative numeric types and in D to support higher order differentiation.

Fields
x: V

The value of the variable.

dx: D

The derivative of the variable.

Implementations

impl<V, D> F<V, D>

pub fn new(x: V, dx: D) -> F<V, D>

Create a new dual number with value x and initial derivative d.

This is equivalent to setting the fields of F directly.

impl<V, D: Zero> F<V, D>

pub fn cst(x: impl Into<V>) -> F<V, D>

Create a new constant.

Use this also to convert from a variable to a constant.

impl<V, D: One> F<V, D>

pub fn var(x: impl Into<V>) -> F<V, D>

Create a new variable.

Use this also to convert from a constant to a variable.

impl<V: Clone, D> F<V, D>

pub fn value(&self) -> V

Get the value of this variable.

impl<V, D: Clone> F<V, D>

pub fn deriv(&self) -> D

Get the current derivative of this variable.

This will be zero if this F is a constant.

impl<V, D> F<V, D>where
Self: Float,

pub fn pow(self, n: impl Into<F<V, D>>) -> F<V, D>

Raise this number to the n’th power.

This is a generic version of Float::powf.